PP-183: Refactor unit tests to different file
